ICFG refactoring stage 1
