+++ lib/WPA/TypeAnalysis.cpp
@@ -44,7 +44,7 @@ llvm::cl::opt<bool> genICFG("genicfg", llvm::cl::init(true), llvm::cl::desc("Gen
 void TypeAnalysis::initialize(SVFModule svfModule) {
     Andersen::initialize(svfModule);
 	if (genICFG) {
-		icfg = new ICFG(ptaCallGraph);
+		icfg = new ICFG();
 		icfg->dump("icfg_initial");
 		icfg->dump("vfg_initial");
 		if (print_stat){
